backend/requirements.txt
# =============================================================================
# AsthmaShield Backend - Complete Dependencies
# =============================================================================

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database & ORM
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0

# Configuration & Environment
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP & API
httpx==0.25.2
requests==2.31.0

# Real-time & WebSockets
websockets==12.0
python-socketio==5.10.0

# Background Tasks & Caching
celery==5.3.4
redis==5.0.1

# File Handling
aiofiles==23.2.1

# AI & Machine Learning
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
tensorflow==2.15.0

# Computer Vision & Image Processing
opencv-python==4.8.1.78
pillow==10.1.0

# Audio Processing
librosa==0.10.1
soundfile==0.12.1

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-xdist==3.3.1
pytest-benchmark==4.0.0
pytest-html==4.1.1
pytest-json-report==1.5.0
factory-boy==3.3.0
faker==20.1.0
requests-mock==1.11.0

# Code Quality & Linting
black==23.11.0
flake8==6.1.0
flake8-docstrings==1.7.0
flake8-import-order==0.18.2
flake8-quotes==3.4.0
mypy==1.7.1
isort==5.12.0
pre-commit==3.5.0
bandit==1.7.5
safety==3.5.0

# Development Tools
ipython==8.17.2
ipdb==0.13.13
jupyter==1.0.0
notebook==7.0.6

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0
myst-parser==0.8.12

# Performance & Profiling
memory-profiler==0.61.0
line-profiler==4.1.2
py-spy==0.3.14

# Database Tools
pgcli==4.0.1

# Monitoring & Logging
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Health Checks
healthcheck==1.3.0

# Rate Limiting
slowapi==0.1.9

# Data Validation
marshmallow==3.20.1

# Date & Time
python-dateutil==2.8.2
pytz==2023.3

# Utilities
click==8.1.7
rich==13.7.0
tqdm==4.66.1
colorama==0.4.6
python-decouple==3.8

# Type Checking
types-requests==2.31.0.10
types-PyYAML==6.0.12.12
types-python-dateutil==2.8.19.14

# Explainable AI
shap==0.42.1
lime==0.2.0.1

# AWS Integration
boto3==1.34.0
